+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+## [1.3.0] - 2023-05-02
+
 ### Added
 
 - RandomSkunk.Results:
@@ -610,7 +612,8 @@ Initial release.
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-[Unreleased]: https://github.com/bfriesen/RandomSkunk.Results/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/bfriesen/RandomSkunk.Results/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/bfriesen/RandomSkunk.Results/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/bfriesen/RandomSkunk.Results/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/bfriesen/RandomSkunk.Results/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/bfriesen/RandomSkunk.Results/compare/v1.0.0-alpha23...v1.0.0